+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `OxidEsales\EshopCommunity\Core\UtilsView:getShopSmartyPluginDirectories()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:_fillCommonSmartyProperties()`
   - `OxidEsales\EshopCommunity\Core\UtilsView:_smartyCompileCheck()`
+  - `OxidEsales\EshopCommunity\Core\Email:_getSmarty()`
+  - `OxidEsales\EshopCommunity\Core\UtilsView:getSmarty()`
   - `oxfunctions:ox_get_template()`
   - `oxfunctions:ox_get_timestamp()`
   - `oxfunctions:ox_get_secure()`